refactor to not remove fields that shouldnt be removed
