+++ src/models/condition.ts
@@ -1,13 +1,13 @@
+import {
+  codeableConceptLabel,
+  findCoding,
+  findCodingByOrderOfPreference,
+  findCodingWithEnrichment,
+} from "@/fhir/codeable-concept";
 import { CONDITION_CODE_PREFERENCE_ORDER } from "@/fhir/conditions";
 import { findReference } from "@/fhir/resource-helper";
 import { ResourceMap } from "@/fhir/types";
 import { compact, uniqWith } from "lodash";
-import {
-  codeableConceptLabel,
-  findCoding,
-  findCodingWithEnrichment,
-  findCodingByOrderOfPreference,
-} from "@/fhir/codeable-concept";
 import { formatDateISOToLocal, formatStringToDate } from "../fhir/formatters";
 import { SYSTEM_CCS, SYSTEM_ICD10, SYSTEM_SNOMED } from "../fhir/system-urls";
 import { PatientModel } from "./patients";
